+++ src/main/java/aQute/bnd/osgi/Builder.java
@@ -238,7 +238,10 @@ public class Builder extends Analyzer {
 			f = new File(f, "MANIFEST.MF");
 		}
 		f.delete();
-		f.getParentFile().mkdirs();
+		File fp = f.getParentFile();
+		if (!fp.exists() && !fp.mkdirs()) {
+			throw new IOException("Could not create directory " + fp);
+		}
 		OutputStream out = new FileOutputStream(f);
 		try {
 			Jar.writeManifest(dot.getManifest(), out);
@@ -281,6 +284,7 @@ public class Builder extends Analyzer {
 	/**
 	 * Answer extra packages. In this case we implement conditional package. Any
 	 */
+	@Override
 	protected Jar getExtra() throws Exception {
 		Parameters conditionals = getParameters(CONDITIONAL_PACKAGE);
 		if (conditionals.isEmpty())
@@ -315,6 +319,7 @@ public class Builder extends Analyzer {
 	 * the setup. We do not want to cleanup if we are going to verify.
 	 */
 
+	@Override
 	public void analyze() throws Exception {
 		super.analyze();
 		cleanupVersion(getImports(), null);
@@ -1078,6 +1083,7 @@ public class Builder extends Analyzer {
 		sourcePath.add(cp);
 	}
 
+	@Override
 	public void close() {
 		super.close();
 	}
